+++ app/class/base64.class.php
@@ -1,0 +1,10 @@
+<?php
+class Base64 {
+    public function encode($input) {
+
+    }
+
+    public function decode($input) {
+
+    }
+}