+++ app/class/controller.class.php
@@ -7,8 +7,12 @@ class Controller {
     }
 
     public function load($filename) {
+        global $template;
         global $notes;
+        global $todos;
         global $user;
+        global $profile;
+        global $bootstrap;
         
         return include $this->path . $filename . '.con.php';
     }