+++ app/class/to_make/installer.class.php
@@ -1,0 +1,19 @@
+<?php
+class Installer {
+
+    public function __construct() {
+        // IF FILE EXIST DIE() AND WARN
+    }
+
+    public function new() {
+
+        // INIT HELE SYSTEEM
+        
+    }
+
+    public function initUser($username) {
+        global $mysql;
+
+        // NIEUWE TABEL VOOR NIEUWE USER
+    }
+}