+++ app/class/views.class.php
@@ -8,10 +8,11 @@ class Views {
 
     public function load($filename) {
         global $template;
-
         global $notes;
         global $todos;
         global $user;
+        global $profile;
+        global $bootstrap;
 
         return include $this->path . $filename . '.view.php';
     }