+++ app/controller/user/sign-out.con.php
@@ -1,1 +1,2 @@
 <?php $user->doSignOut(); ?>
+<?php header("Location: index.php"); ?>