+++ app/views/profile/password.view.php
@@ -1,0 +1,2 @@
+<?php echo $template->addTitle("Password Change"); ?>
+<?php echo $profile->passwordChangeForm(); ?>