+++ autoload.php
@@ -1,14 +1,5 @@
 <?php
-// Voor debugging aan in live modus uit zetten.
-$debug          = true;
-
-// MySQL gegevens
-$sqlHost        = "mysql";
-$sqlUsername    = "root";
-$sqlPassword    = "root";
-$sqlDatabase    = "notes";
-
-$lang = "en";
+include "config.php";
 
 if ($debug === true) {
     ini_set('display_errors', 1);
@@ -18,12 +9,15 @@ if ($debug === true) {
 
 foreach (glob("app/class/*.php") as $file) { include $file; }
 
-$mysql = new mysqli($sqlHost, $sqlUsername, $sqlPassword, $sqlDatabase);
-$template = new Template("app/assets", $lang);
-$views = new Views();
-$controller = new Controller();
-
-$notes = new Notes();
-$profile = new Profile();
-$user = new User();
+$mysql          = new mysqli($sqlHost, $sqlUsername, $sqlPassword, $sqlDatabase);
+$template       = new Template("app/assets", $lang);
+$views          = new Views();
+$controller     = new Controller();
+$profile        = new Profile();
+$notes          = new Notes();
+$profile        = new Profile();
+$user           = new User();
+$bootstrap      = new Bootstrap();
+$dateTime       = new DT($lang);
+$baseSixtyFour  = new Base64();
 ?>