+++ config.php
@@ -1,0 +1,9 @@
+<?php
+$debug = true;
+
+$sqlHost        = "mysql";
+$sqlUsername    = "root";
+$sqlPassword    = "root";
+$sqlDatabase    = "notes";
+
+$lang = "en";