+++ sign-out.php
@@ -1,0 +1,6 @@
+<?php include "autoload.php"; ?>
+<?php session_start(); ?>
+<?php 
+$controller->load("user/sign-out");
+$views->load("user/sign-out");
+?>