+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Version: 0.0.2
+- ADDED: Logger function that registers fault logins in the class "User"
+- ADDED: Label right-top with the name of the current user
+- ADDED: code is prepared for building a admin-console around it
+- ADDED: Password change
+- FIX: Rewrited the code with function name "addJS" & "addCSS", now a boolean is required to let the script know if its a local file or a file hosted on a CDN
+- FIX: Login verify issues. Added two new pages "sign-in.php" and "sign-out.php"
+
 ## Version: 0.0.1
 - Init